Disabled styles
This uses `graytext` as the color keyword for both the text color and SVG arrow fill. It also removes the border color change on hover, which can otherwise indicate it is still interactive.

Because many accessibility best practices rely on `aria-disabled` I also added that as a selector. If nothing else, it is a visual reminder to developers who forget to toggle `aria-disabled` when they set `disabled`.

Tested in Firefox, Edge, Internet Explorer and Chrome all on Windows. No Mac handy today to test.